# bot-api/openapi.yaml
openapi: 3.0.3
info:
  title: LendWizely AI Bot API
  description: REST API for SMS outreach, intake, bank parsing, offer generation, background checks, and e-sign
  version: 1.0.0
  contact:
    name: LendWizely Team
    email: dev@lendwizely.com
servers:
  - url: https://api.lendwizely.com
    description: Production server
  - url: http://localhost:8080
    description: Development server
security:
  - BearerAuth: []
paths:
...
  /plaid/link-token:
    post:
      summary: Create Plaid link token
      description: Generate link token for Plaid Link
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - client_id
              properties:
                client_id:
                  type: string
      responses:
        '200':
          description: Link token created
          content:
            application/json:
              schema:
                type: object
                properties:
                  link_token:
                    type: string
                  expiration:
                    type: string
                    format: date-time
  /plaid/exchange:
    post:
      summary: Exchange public token
      description: Exchange Plaid public token for access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - public_token
                - client_id
              properties:
                public_token:
                  type: string
                client_id:
                  type: string
      responses:
        '200':
          description: Token exchanged successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                  item_id:
                    type: string
  /plaid/transactions:
    get:
      summary: Get transactions
      description: Retrieve transactions for date range
      parameters:
        - name: client_id
          in: query
          required: true
          schema:
            type: string
        - name: start_date
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: end_date
          in: query
          required: true
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Transactions retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  transactions:
                    type: array
                    items:
                      $ref: '#/components/schemas/Transaction'
  /plaid/statements:
    get:
      summary: List statements
      description: Get available bank statements
      parameters:
        - name: client_id
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Statements list
          content:
            application/json:
              schema:
                type: object
                properties:
                  statements:
                    type: array
                    items:
                      $ref: '#/components/schemas/Statement'
  /plaid/statements/{statement_id}/download:
    get:
      summary: Download statement
      description: Download statement PDF
      parameters:
        - name: statement_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Statement PDF
          content:
            application/pdf:
              schema:
                type: string
                format: binary
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Transaction:
      type: object
      properties:
        id:
          type: string
        amount:
          type: number
        date:
          type: string
          format: date
        description:
          type: string
        account_id:
          type: string
    Statement:
      type: object
      properties:
        id:
          type: string
        account_id:
          type: string
        statement_date:
          type: string
          format: date
        file_size:
          type: integer
...